+++ app-server/src/main/java/com/inventory/app/server/service/media/TelevisionService.java
@@ -2,8 +2,11 @@ package com.inventory.app.server.service.media;
 
 import com.inventory.app.server.config.MediaInventoryAdditionalAttributes;
 import com.inventory.app.server.entity.media.TelevisionShow;
+import com.inventory.app.server.error.NoChangesToUpdateException;
+import com.inventory.app.server.error.ResourceAlreadyExistsException;
+import com.inventory.app.server.error.ResourceNotFoundException;
 import com.inventory.app.server.repository.IBaseDao;
-import com.inventory.app.server.utility.RestPreConditions;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,49 +26,77 @@ public class TelevisionService {
     }
 
     public List<TelevisionShow> getAllBooksByCollectionTitle(String collectionTitle) {
-        return dao.findByField("collection_name", collectionTitle);
+        List<TelevisionShow> televisionShowList = dao.findByField("collection_name", collectionTitle);
+        if (televisionShowList.isEmpty()) {
+            throw new ResourceNotFoundException("No television show results found with collection title " + collectionTitle);
+        }
+        return televisionShowList;
     }
 
-    public List<TelevisionShow> getAllTelevisionShowsByWriter(List<String> writer) {
-        return dao.findByField(MediaInventoryAdditionalAttributes.WRITERS.getJsonKey(), writer);
+    public List<TelevisionShow> getAllTelevisionShowsByEpisode(List<String> episodes) {
+        List<TelevisionShow> televisionShowList = dao.findByField(MediaInventoryAdditionalAttributes.EPISODES.getJsonKey(), episodes);
+        if (televisionShowList.isEmpty()) {
+            throw new ResourceNotFoundException("No television show results found with episode(s) " + episodes);
+        }
+        return televisionShowList;
     }
 
     public List<TelevisionShow> getAllTelevisionShowsByGenre(String genre) {
-        return dao.findByField("genre", genre);
+        List<TelevisionShow> televisionShowList = dao.findByField("genre", genre);
+        if (televisionShowList.isEmpty()) {
+            throw new ResourceNotFoundException("No television show results found for genre " + genre);
+        }
+        return televisionShowList;
     }
 
-    public List<TelevisionShow> getAllTelevisionShows() {
-        return dao.findAll();
+    public List<TelevisionShow> getAll() {
+        List<TelevisionShow> televisionShowList = dao.findAll();
+        if (televisionShowList.isEmpty()) {
+            throw new ResourceNotFoundException("No television show data exists.");
+        }
+        return televisionShowList;
     }
 
-    public TelevisionShow getTelevisionShowById(Long id) {
-        return dao.findOne(id);
+    public TelevisionShow getById(Long id) {
+        try {
+            return dao.findOne(id);
+        } catch (Exception e) {
+            if ( e.getClass().isInstance(EntityNotFoundException.class)) {
+                throw new ResourceNotFoundException("No book exists with id: " + id);
+            } else {
+                throw e;
+            }
+        }
     }
 
     public TelevisionShow create(TelevisionShow televisionShow) {
-        // validations before performing create
-        RestPreConditions.checkAlreadyExists(televisionShowAlreadyExists(televisionShow), televisionShow);
-
+        if(televisionShowAlreadyExists(televisionShow)) {
+            throw new ResourceAlreadyExistsException("Cannot create television show because television show already exists: " + televisionShow);
+        }
         TelevisionShow televisionShowToSave = cloneTelevisionShow(televisionShow);
         televisionShowToSave.setId(null);
         televisionShowToSave.setVersion(1);
-
         return dao.createOrUpdate(televisionShowToSave);
     }
 
     public TelevisionShow update(TelevisionShow updatedTelevisionShow) {
-        //validations before performing the update
-        TelevisionShow existingTelevisionShow = RestPreConditions.checkFound(getTelevisionShowById(updatedTelevisionShow.getId()));
-        RestPreConditions.checkEquals(existingTelevisionShow, updatedTelevisionShow);
-
+        if (!televisionShowAlreadyExists(updatedTelevisionShow)) {
+            throw new ResourceNotFoundException("Cannot update television show because television show does not exist: " + updatedTelevisionShow);
+        }
+        TelevisionShow existingTelevisionShow = getById(updatedTelevisionShow.getId());
+        if (verifyIfTelevisionShowUpdated(existingTelevisionShow, updatedTelevisionShow)) {
+            throw new NoChangesToUpdateException("No updates in television show to save. Will not proceed with update. Existing Television Show: " + existingTelevisionShow + "Updated Television Show: " + updatedTelevisionShow);
+        }
         updatedTelevisionShow = cloneTelevisionShow(existingTelevisionShow, updatedTelevisionShow);
         updatedTelevisionShow.setVersion(existingTelevisionShow.getVersion() + 1);
-
         return dao.createOrUpdate(updatedTelevisionShow);
     }
 
     public TelevisionShow deleteById(Long id){
-        TelevisionShow televisionShow = RestPreConditions.checkFound(getTelevisionShowById(id));
+        TelevisionShow televisionShow = getById(id);
+        if (televisionShow == null) {
+            throw new ResourceNotFoundException("Cannot delete television show because television show does not exist.");
+        }
         dao.deleteById(id);
         return televisionShow;
     }
@@ -82,8 +113,12 @@ public class TelevisionService {
     }
 
     private boolean televisionShowAlreadyExists(TelevisionShow televisionShow) {
-        return getAllTelevisionShowsByWriter(televisionShow.getWriters())
+        return getAllTelevisionShowsByEpisode(televisionShow.getEpisodes())
                 .stream()
                 .anyMatch(t -> televisionShow.getTitle().equals(t.getTitle()));
+    }
+
+    private boolean verifyIfTelevisionShowUpdated(TelevisionShow existingTelevisionShow, TelevisionShow updatedTelevisionShow) {
+        return existingTelevisionShow.equals(updatedTelevisionShow);
     }
 }
